+++ extensions/youtube-downloader/CHANGELOG.md
@@ -1,5 +1,9 @@
 # YouTube Downloader Changelog
 
+## [Insert active tab URL] - 2025-02-02
+
+- If the raycast browser extension is installed, the extension will automatically insert the active tab URL into the input field
+
 ## [Improve URL Validator] - 2025-01-23
 
 - Improve `isYouTubeURL` function